--- a/Project.toml
+++ b/Project.toml
@@ -12,6 +12,7 @@ LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
 MathOptInterface = "b8f27783-ece8-5eb3-8dc8-9495eed66fee"
 MosekTools = "1ec41992-ff65-5c91-ac43-2df89e9693a4"
 MultivariatePolynomials = "102ac46a-7ee4-5c85-9060-abc95bfdeaa3"
+MutableArithmetics = "d8a4904e-b15c-11e9-3269-09a3773c0cb0"
 OrderedCollections = "bac558e1-5e72-5ebc-8fee-abe8a469f55d"
 PolyJuMP = "ddf597a6-d67e-5340-b84c-e37d84115374"
 Reexport = "189a3867-3050-52da-a836-e630ba90ab69"
@@ -23,13 +24,14 @@ DynamicPolynomials = "~0.3.2"
 JuMP = "~0.20"
 MathOptInterface = "~0.9.1"
 MultivariatePolynomials = "~0.3.1"
+MutableArithmetics = "0.1"
 PolyJuMP = "~0.3"
 SemialgebraicSets = "~0.2"
 SumOfSquares = "~0.3"
 
 [extras]
-Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 CSDP = "0a46da34-8e4b-519e-b418-48813639ff34"
+Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 
 [targets]
 test = ["CSDP"]