--- a/src/relax.jl
+++ b/src/relax.jl
@@ -1,40 +1,5 @@
 export relax!, termination_status
 
-function add_prod!(p::AbstractPolynomial{JuMP.AffExpr}, q::AbstractPolynomialLike{<:Number}, args...)
-    pt = terms(p)
-    qt = terms(q)
-    pelst = iterate(pt)
-    qelst = iterate(qt)
-    r = zero(q)
-    while qelst !== nothing
-        qel, qst = qelst
-        if pelst === nothing
-            r += qel
-            qelst = iterate(qt, qst)
-        else
-            pel, pst = pelst
-            if monomial(pel) == monomial(qel)
-                JuMP.add_to_expression!(coefficient(pel), coefficient(qel), args...)
-                qelst = iterate(qt, qst)
-                pelst = iterate(pt, pst)
-            elseif monomial(pel) < monomial(qel)
-                r += qel
-                qelst = iterate(qt, qst)
-            else
-                pelst = iterate(pt, pst)
-            end
-        end
-    end
-    if iszero(nterms(r))
-        return p
-    else
-        return p + *(r, args...)
-    end
-end
-function add_prod!(p::AbstractPolynomial{JuMP.AffExpr}, q::Number, args...)
-    add_prod!(p, constantterm(q, p), args...)
-end
-
 function dual_variable(dual::JuMP.Model, momcon::MomCon, sense::MOI.OptimizationSense)
     if momcon.set isa MOI.EqualTo
         variable = @variable(dual)
@@ -79,7 +44,7 @@ function relax!(gmp::GMPModel, order::Int, optimizer::OptimizerFactory)
 
     for (momcon, variable) in zip(gmp.constraints, gmp.dref)
         for meas in measures(momcon)
-            dlhs[meas] = add_prod!(dlhs[meas], momcon.func.momdict[meas], variable)
+            dlhs[meas] = MA.add_mul!(dlhs[meas], momcon.func.momdict[meas], variable)
         end
     end
 
@@ -94,7 +59,7 @@ function relax!(gmp::GMPModel, order::Int, optimizer::OptimizerFactory)
                 obj = -obj
             end
             # Modifies dlhs[meas] but it's ok since it not used anywhere else
-            p = add_prod!(p, obj)
+            p = MA.add_mul!(p, obj)
         end
         gmp.cref[meas] = @constraint(gmp.dual, p in meas.cert, domain = meas.supp, maxdegree = 2*order)
     end