On the burgers benchmark added in https://github.com/lanl-ansi/MomentOpt.jl/pull/5, here are the timings before and after MutableArithmetics (MA) in seconds:

| Order | Before MA [s] | After MA [s] |
|-------|---------------|--------------|
| 1     | 0.126383      | 0.014443     |
| 2     | 0.132074      | 0.022544     |
| 3     | 0.138352      | 0.035078     |
| 4     | 0.141458      | 0.041759     |
| 5     | 0.226350      | 0.054297     |
| 6     | 0.440967      | 0.076973     |
| 7     | 1.127554      | 0.155279     |
| 8     | 3.088487      | 0.240168     |
| 9     | 5.016221      | 0.411574     |
| 10    | 11.754727     | 0.670018     |
| 11    | 20.357476     | 1.386744     |
| 12    | 44.614137     | 2.059797     |

We can see that this also goes with a large reduction of memory allocation:
Before MA:
```julia
julia> burgers(12)
 44.614137 seconds (13.44 M allocations: 17.344 GiB, 5.85% gc time)
```
After MA:
```julia
julia> burgers(12)
  2.059797 seconds (6.82 M allocations: 1.043 GiB, 20.51% gc time)
```

Most of the performance improvements come from https://github.com/JuliaOpt/SumOfSquares.jl/pull/130, https://github.com/JuliaAlgebra/DynamicPolynomials.jl/pull/52 and https://github.com/JuliaOpt/MathOptInterface.jl/pull/924