@tweisser 2019-12-12
Nice :+1: 
I'm sure you are already looking at the compatibility problem causing travis to fail building...