@blegat 2019-12-13
We need to wait 2 days for MA to get registered. Then we'll release MOI and MultivariatePolynomials, ... Then this PR will work :P